+++ openrem/remapp/exports/rf_export.py
@@ -1,4 +1,3 @@
-# This Python file uses the following encoding: utf-8
 #    OpenREM - Radiation Exposure Monitoring tools for the physicist
 #    Copyright (C) 2012,2013  The Royal Marsden NHS Foundation Trust
 #
@@ -29,12 +28,19 @@
 
 """
 
+import datetime
 import logging
+
 from celery import shared_task
 from django.core.exceptions import ObjectDoesNotExist
-from ..exports.export_common import text_and_date_formats, common_headers, generate_sheets, sheet_name, \
-    get_common_data, get_xray_filter_info, create_xlsx, create_csv, write_export, create_summary_sheet, \
-    get_pulse_data, abort_if_zero_studies
+from django.db.models import Avg, Max, Min
+
+from ..exports.export_common import (text_and_date_formats, common_headers, generate_sheets, sheet_name,
+                                     get_common_data, get_xray_filter_info, create_xlsx, create_csv, write_export,
+                                     create_summary_sheet, get_pulse_data, abort_if_zero_studies, create_export_task,
+                                     get_patient_study_data)
+from ..interface.mod_filters import RFSummaryListFilter, RFFilterPlusPid
+from ..models import GeneralStudyModuleAttr, IrradEventXRayData
 from ..tools.get_values import return_for_export
 
 logger = logging.getLogger(__name__)
@@ -218,27 +224,10 @@ def rfxlsx(filterdict, pid=False, name=None, patid=None, user=None):
     :return: Saves xlsx file into Media directory for user to download
     """
 
-    import datetime
-    from django.db.models import Max, Min, Avg
-    from ..models import GeneralStudyModuleAttr, IrradEventXRayData
-    from ..models import Exports
-    from ..interface.mod_filters import RFSummaryListFilter, RFFilterPlusPid
-    import uuid
-
-    tsk = Exports.objects.create()
-
-    tsk.task_id = rfxlsx.request.id
-    if tsk.task_id is None:  # Required when testing without celery
-        tsk.task_id = u'NotCelery-{0}'.format(uuid.uuid4())
-    tsk.modality = u"RF"
-    tsk.export_type = u"XLSX export"
     datestamp = datetime.datetime.now()
-    tsk.export_date = datestamp
-    tsk.progress = u'Query filters imported, task started'
-    tsk.status = u'CURRENT'
-    tsk.includes_pid = bool(pid and (name or patid))
-    tsk.export_user_id = user
-    tsk.save()
+    tsk = create_export_task(celery_uuid=rfxlsx.request.id, modality='RF', export_type='XLSX export',
+                             date_stamp=datestamp, pid=bool(pid and (name or patid)), user=user,
+                             filters_dict=filterdict)
 
     tmpxlsx, book = create_xlsx(tsk)
     if not tmpxlsx:
@@ -531,23 +520,10 @@ def exportFL2excel(filterdict, pid=False, name=None, patid=None, user=None):
     :return: Saves csv file into Media directory for user to download
     """
 
-    import datetime
-    from ..models import GeneralStudyModuleAttr
-    from ..models import Exports
-    from ..interface.mod_filters import RFSummaryListFilter, RFFilterPlusPid
-
-    tsk = Exports.objects.create()
-
-    tsk.task_id = exportFL2excel.request.id
-    tsk.modality = u"RF"
-    tsk.export_type = u"CSV export"
     datestamp = datetime.datetime.now()
-    tsk.export_date = datestamp
-    tsk.progress = u'Query filters imported, task started'
-    tsk.status = u'CURRENT'
-    tsk.includes_pid = bool(pid and (name or patid))
-    tsk.export_user_id = user
-    tsk.save()
+    tsk = create_export_task(celery_uuid=exportFL2excel.request.id, modality='RF', export_type='CSV export',
+                             date_stamp=datestamp, pid=bool(pid and (name or patid)), user=user,
+                             filters_dict=filterdict)
 
     tmpfile, writer = create_csv(tsk)
     if not tmpfile:
@@ -614,27 +590,17 @@ def exportFL2excel(filterdict, pid=False, name=None, patid=None, user=None):
 
 @shared_task
 def rfopenskin(studyid):
-    u"""Export filtered RF database data to multi-sheet Microsoft XSLX files.
+    """Export single RF study data to OpenSkin RF csv sheet.
 
     :param studyid: RF study database ID.
     :type studyid: int
 
-    u"""
+    """
 
-    import datetime
-    from ..models import GeneralStudyModuleAttr
-    from ..models import Exports
-
-    tsk = Exports.objects.create()
-
-    tsk.task_id = rfopenskin.request.id
-    tsk.modality = "RF-OpenSkin"
-    tsk.export_type = "OpenSkin RF csv export"
     datestamp = datetime.datetime.now()
-    tsk.export_date = datestamp
-    tsk.progress = u'Query filters imported, task started'
-    tsk.status = u'CURRENT'
-    tsk.save()
+    tsk = create_export_task(celery_uuid=rfopenskin.request.id, modality='RF-OpenSkin',
+                             export_type='OpenSkin RF csv export', date_stamp=datestamp,
+                             pid=False, user=None, filters_dict={'study_id': studyid})
 
     tmpfile, writer = create_csv(tsk)
     if not tmpfile:
@@ -826,25 +792,9 @@ def rf_phe_2019(filterdict, user=None):
     :return: Saves Excel file into media directory for user to download
     """
 
-    import datetime
-    import uuid
-    from django.db.models import Max, Min
-    from ..exports.export_common import get_patient_study_data
-    from ..models import Exports, GeneralStudyModuleAttr, IrradEventXRayData
-    from ..interface.mod_filters import RFSummaryListFilter
-
-    tsk = Exports.objects.create()
-    tsk.task_id = rf_phe_2019.request.id
-    if tsk.task_id is None:  # Required when testing without celery
-        tsk.task_id = u'NotCelery-{0}'.format(uuid.uuid4())
-    tsk.modality = u"RF"
-    tsk.export_type = u"PHE RF 2019 export"
     datestamp = datetime.datetime.now()
-    tsk.export_date = datestamp
-    tsk.progress = u'Query filters imported, task started'
-    tsk.status = u'CURRENT'
-    tsk.export_user_id = user
-    tsk.save()
+    tsk = create_export_task(celery_uuid=rf_phe_2019.request.id, modality='RF', export_type='PHE RF 2019 export',
+                             date_stamp=datestamp, pid=False, user=user, filters_dict=filterdict)
 
     tmp_xlsx, book = create_xlsx(tsk)
     if not tmp_xlsx:
@@ -928,17 +878,22 @@ def rf_phe_2019(filterdict, user=None):
         fluoro_events = events.exclude(
             irradiation_event_type__code_value__contains='11361')  # acq events are 113611, 113612, 113613
         acquisition_events = events.filter(irradiation_event_type__code_value__contains='11361')
-        row_data += [
-            u' | '.join(fluoro_events.order_by().values_list('acquisition_protocol', flat=True).distinct()),
-            u' | '.join(fluoro_events.order_by().values_list(
-                'irradeventxraysourcedata__fluoro_mode__code_meaning', flat=True).distinct()),
-        ]
+        try:
+            row_data += [' | '.join(
+                fluoro_events.order_by().values_list('acquisition_protocol', flat=True).distinct()), ]
+        except TypeError:
+            row_data += ['', ]
+        try:
+            row_data += [' | '.join(fluoro_events.order_by().values_list(
+                'irradeventxraysourcedata__fluoro_mode__code_meaning', flat=True).distinct()), ]
+        except TypeError:
+            row_data += ['', ]
         fluoro_frame_rates = fluoro_events.order_by().values_list(
             'irradeventxraysourcedata__pulse_rate', flat=True).distinct()
         column_aq = ''
         if len(fluoro_frame_rates) > 1:
             column_aq += u'Fluoro: '
-            column_aq += u' | '.join(format(x, "1.1f") for x in fluoro_frame_rates)
+            column_aq += u' | '.join(format(x, "1.1f") for x in fluoro_frame_rates if x is not None)
             column_aq += u' fps. '
             row_data += [
                 u'Multiple rates',